Split models into a file for each model
